pyproject.toml
[project]
name = "wsn-clustering-benchmark"
version = "1.0.0"
description = "Simulation framework for benchmarking WSN clustering protocols with explicit control-plane energy modeling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.urls]
Repository = "https://github.com/Hzz-Git/WSN-Clustering-Benchmark"
